feat: deploy service with aws cdk
