+++ productService/.npmignore
