+++ productService/lambda/mock/products.js
@@ -1,0 +1,19 @@
+const products = [
+    {
+        "id": 1,
+        "name": "Iphone 14",
+        "price": 100
+    },
+    {
+        "id": 2,
+        "name": "Apple whatch",
+        "price": 200
+    },
+    {
+        "id": 3,
+        "name": "IPhone headphones",
+        "price": 300
+    }
+];
+
+exports.products = products;
